+++ openai_function_call/dsl/__init__.py
@@ -1,0 +1,5 @@
+from .completion import ChatCompletion
+from .multitask import MultiTask
+from .messages import *
+
+__all__ = ["ChatCompletion", "MultiTask", "messages"]
