+++ tests/test_messages.py
@@ -1,9 +1,9 @@
-from prompt_builder import *
+from openai_function_call.dsl import messages as m
 
 
 def test_create_message():
-    assert Message(
-        role=MessageRole.SYSTEM,
+    assert m.Message(
+        role=m.MessageRole.SYSTEM,
         content="Hello, world!",
     ).dict() == {
         "role": "system",
@@ -12,7 +12,7 @@ def test_create_message():
 
 
 def test_create_user_message():
-    assert UserMessage(
+    assert m.UserMessage(
         content="Hello, world!",
     ).dict() == {
         "role": "user",
@@ -21,35 +21,35 @@ def test_create_user_message():
 
 
 def test_create_system_message():
-    assert SystemMessage(content="I am nice").dict() == {
+    assert m.SystemMessage(content="I am nice").dict() == {
         "role": "system",
         "content": "I am nice",
     }
 
 
 def test_assistance_message():
-    assert AssistantMessage(content="I am nice").dict() == {
+    assert m.AssistantMessage(content="I am nice").dict() == {
         "role": "assistant",
         "content": "I am nice",
     }
 
 
 def test_create_tagged_message():
-    assert TaggedMessage(content="I am nice", tag="data").dict() == {
+    assert m.TaggedMessage(content="I am nice", tag="data").dict() == {
         "role": "user",
         "content": "<data>I am nice</data>",
     }
 
 
 def test_expert_system_message():
-    assert ExpertSystem(task="task").dict() == {
+    assert m.ExpertSystem(task="task").dict() == {
         "role": "system",
         "content": "You are a world class, state of the art agent capable of correctly completing the task: `task`",
     }
 
 
 def test_chain_of_thought_message():
-    assert ChainOfThought().dict() == {
+    assert m.ChainOfThought().dict() == {
         "role": "assistant",
         "content": "Lets think step by step to get the correct answer:",
     }
